requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
Flask>=2.2.0